ajustadas as classes layer
